+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/async/AsyncResultsReadySpec.groovy
@@ -66,7 +66,7 @@ class AsyncResultsReadySpec extends AsyncFunctionalSpec {
     */
 
     static final String QUERY =
-            "http://localhost:9998/data/shapes/day?dateTime=2016-08-30%2F2016-08-31&metrics=height&asyncAfter=0"
+            "http://localhost:9998/data/shapes/day?dateTime=2016-08-30%2F2016-08-31&metrics=height&asyncAfter=always"
 
     final CountDownLatch jobMetadataReady = new CountDownLatch(1)
 
@@ -131,7 +131,7 @@ class AsyncResultsReadySpec extends AsyncFunctionalSpec {
         [
                 data: {[
                         metrics: ["height"],
-                        asyncAfter: ["0"],
+                        asyncAfter: ["always"],
                         dateTime: ["2016-08-30/2016-08-31"]
                 ]},
                 syncResults: {